words_extractor_py/requirements.txt
#PyICU>=2.7.4
PyYAML>=5.4.1